+++ scenarios/argoverse/vehicle_following/ff6dc43b-dd27-4fe4-94b6-5c1b3940daed_agents_1/scenario.py
@@ -15,6 +15,7 @@ traffic_histories = [
     )
 ]
 
+duration = 11
 ego_mission = [t.EndlessMission(begin=("road-243772234-243772033", 1, 10.9))]
 
 leader_id = "history-vehicle-40031$"
@@ -24,7 +25,12 @@ gen_scenario(
         ego_missions=ego_mission,
         map_spec=t.MapSpec(source=f"{scenario_path}", lanepoint_spacing=1.0),
         traffic_histories=traffic_histories,
-        scenario_metadata=t.ScenarioMetadata(leader_id, Colors.Blue),
+        scenario_metadata=t.ScenarioMetadata(
+            actor_of_interest_re_filter=leader_id,
+            actor_of_interest_color=Colors.Blue,
+            scenario_difficulty=0.3,
+            scenario_duration=duration,
+        ),
     ),
     output_dir=Path(__file__).parent,
 )