+++ scenarios/sumo/intersections/1_to_3lane_left_turn_sumo_c_agents_1/scenario.py
@@ -24,7 +24,6 @@ from pathlib import Path
 
 from smarts.sstudio import gen_scenario
 from smarts.sstudio.types import (
-    Distribution,
     Flow,
     Mission,
     Route,
@@ -36,67 +35,81 @@ from smarts.sstudio.types import (
 
 normal = TrafficActor(
     name="car",
-    speed=Distribution(sigma=0.3, mean=1.0),
 )
 
-# flow_name = (start_lane, end_lane,)
-route_opt = [
-    (0, 0),
-    (1, 1),
+vertical_routes = [
+    ("E2", 0, "E7", 0),
+    ("E8", 0, "E1", 1),
 ]
 
-# Traffic combinations = 3C2 + 3C3 = 3 + 1 = 4
-# Repeated traffic combinations = 4 * 100 = 400
-min_flows = 2
-max_flows = 2
-route_comb = [
-    com
-    for elems in range(min_flows, max_flows + 1)
-    for com in combinations(route_opt, elems)
-] * 100
+horizontal_routes = [
+    ("E3", 0, "E5", 0),
+    ("E3", 1, "E5", 1),
+    ("E3", 2, "E5", 2),
+    ("E6", 1, "E4", 1),
+    ("E6", 0, "E4", 0),
+]
 
+turn_left_routes = [
+    ("E8", 0, "E5", 2),
+    ("E6", 1, "E1", 1),
+    ("E2", 1, "E4", 1),
+    ("E3", 2, "E7", 0),
+]
+
+turn_right_routes = [
+    ("E6", 0, "E7", 0),
+    ("E3", 0, "E1", 0),
+    ("E2", 0, "E5", 0),
+    ("E8", 0, "E4", 0),
+]
+
+# Total route combinations = 14C1 + 14C2 + 14C3 + 14C4 = 1470
+all_routes = vertical_routes + horizontal_routes + turn_left_routes + turn_right_routes
+route_comb = [com for elems in range(4, 5) for com in combinations(all_routes, elems)]
 traffic = {}
 for name, routes in enumerate(route_comb):
     traffic[str(name)] = Traffic(
         flows=[
             Flow(
                 route=Route(
-                    begin=("gneE3", r[0], 0),
-                    end=("gneE3", r[1], "max"),
+                    begin=(f"{r[0]}", r[1], 0),
+                    end=(f"{r[2]}", r[3], "max"),
                 ),
                 # Random flow rate, between x and y vehicles per minute.
-                rate=60 * random.uniform(5, 10),
+                rate=60 * random.uniform(2, 5),
                 # Random flow start time, between x and y seconds.
-                begin=random.uniform(0, 5),
+                begin=random.uniform(0, 3),
                 # For an episode with maximum_episode_steps=3000 and step
                 # time=0.1s, maximum episode time=300s. Hence, traffic set to
                 # end at 900s, which is greater than maximum episode time of
                 # 300s.
                 end=60 * 15,
                 actors={normal: 1},
-                randomly_spaced=True,
             )
             for r in routes
         ]
     )
 
+route = Route(begin=("E8", 0, 5), end=("E5", 0, "max"))
+
 default_speed = 13
-route_length = 200
+route_length = 100
 duration = (route_length / default_speed) * 2
-route = Route(begin=("gneE3", 0, 5), end=("gneE3", 1, "max"))
+
 ego_missions = [
     Mission(
         route=route,
-        start_time=17,  # Delayed start, to ensure road has prior traffic.
+        start_time=15,  # Delayed start, to ensure road has prior traffic.
     )
 ]
-# runtime = 17
+
 gen_scenario(
     scenario=Scenario(
         traffic=traffic,
         ego_missions=ego_missions,
         scenario_metadata=ScenarioMetadata(
-            scenario_difficulty=0.3,
+            scenario_difficulty=0.6,
             scenario_duration=duration,
         ),
     ),