+++ scenarios/sumo/vehicle_following/merge_exit_sumo_t_agents_2/scenario.py
@@ -110,7 +110,11 @@ ego_missions = [
         ),
     ),
 ]
+default_speed = 13
+route_length = 600
+duration = (route_length / default_speed) * 2
 
+leader_id = "Leader-007"
 gen_scenario(
     scenario=Scenario(
         traffic=traffic,
@@ -118,7 +122,12 @@ gen_scenario(
         map_spec=MapSpec(
             source=Path(__file__).resolve().parents[0], lanepoint_spacing=1.0
         ),
-        scenario_metadata=ScenarioMetadata("Leader-007", Colors.Blue),
+        scenario_metadata=ScenarioMetadata(
+            actor_of_interest_re_filter=leader_id,
+            actor_of_interest_color=Colors.Blue,
+            scenario_difficulty=3,
+            scenario_duration=duration,
+        ),
     ),
     output_dir=Path(__file__).parent,
 )