+++ smarts/benchmark/driving_smarts/v2023/metric_formula_drive.py
@@ -25,13 +25,27 @@ from typing import Dict
 import numpy as np
 
 from smarts.env.gymnasium.wrappers.metric.costs import Costs
-from smarts.env.gymnasium.wrappers.metric.formula import FormulaBase, Score, avg_costs
+from smarts.env.gymnasium.wrappers.metric.formula import (
+    FormulaBase,
+    Score,
+    agent_scores,
+    agent_weights,
+    score_rule_violation,
+    weighted_score,
+)
 from smarts.env.gymnasium.wrappers.metric.params import (
+    Collisions,
     Comfort,
     DistToDestination,
     DistToObstacles,
     JerkLinear,
+    LaneCenterOffset,
+    OffRoad,
     Params,
+    SpeedLimit,
+    Steps,
+    VehicleGap,
+    WrongWay,
 )
 from smarts.env.gymnasium.wrappers.metric.types import Record
 
@@ -51,75 +65,83 @@ class Formula(FormulaBase):
             Params: Cost function parameters.
         """
         params = Params(
-            comfort=Comfort(
-                active=True,
-            ),
-            dist_to_destination=DistToDestination(
-                active=True,
-            ),
-            dist_to_obstacles=DistToObstacles(
-                active=False,
-            ),
+            collisions=Collisions(active=False),
+            comfort=Comfort(active=True),
+            dist_to_destination=DistToDestination(active=True),
+            dist_to_obstacles=DistToObstacles(active=False),
             jerk_linear=JerkLinear(active=False),
+            lane_center_offset=LaneCenterOffset(active=True),
+            off_road=OffRoad(active=False),
+            speed_limit=SpeedLimit(active=True),
+            steps=Steps(active=True),
+            vehicle_gap=VehicleGap(active=False),
+            wrong_way=WrongWay(active=True),
         )
         return params
 
-    def score(self, records_sum: Dict[str, Dict[str, Record]]) -> Score:
+    def score(self, records: Dict[str, Dict[str, Record]]) -> Score:
         """
         Computes several sub-component scores and one total combined score named
         "Overall" on the wrapped environment.
 
-        +-------------------+--------+-----------------------------------------------------------+
-        |                   | Range  | Remarks                                                   |
-        +===================+========+===========================================================+
-        | Overall           | [0, 1] | Total score. The higher, the better.                      |
-        +-------------------+--------+-----------------------------------------------------------+
-        | DistToDestination | [0, 1] | Remaining distance to destination. The lower, the better. |
-        +-------------------+--------+-----------------------------------------------------------+
-        | Time              | [0, 1] | Time taken to complete scenario. The lower, the better.   |
-        +-------------------+--------+-----------------------------------------------------------+
-        | HumannessError    | [0, 1] | Humanness indicator. The lower, the better.               |
-        +-------------------+--------+-----------------------------------------------------------+
-        | RuleViolation     | [0, 1] | Traffic rules compliance. The lower, the better.          |
-        +-------------------+--------+-----------------------------------------------------------+
+        Args:
+            records (Dict[str, Dict[str, Record]]): Records.
 
         Returns:
-            Score: Contains "Overall", "DistToDestination", "VehicleGap",
-            "HumannessError", and "RuleViolation" scores.
+            Score: "Overall" score and other sub-component scores.
         """
 
-        costs_final = avg_costs(records_sum=records_sum)
-
-        # Compute sub-components of score.
-        dist_to_destination = costs_final.dist_to_destination
-        humanness_error = _humanness_error(costs=costs_final)
-        rule_violation = _rule_violation(costs=costs_final)
-        time = costs_final.steps
-        overall = (
-            0.25 * (1 - dist_to_destination)
-            + 0.25 * (1 - time)
-            + 0.25 * (1 - humanness_error)
-            + 0.25 * (1 - rule_violation)
-        )
-
-        return Score(
-            {
-                "overall": overall,
-                "dist_to_destination": dist_to_destination,
-                "time": time,
-                "humanness_error": humanness_error,
-                "rule_violation": rule_violation,
-            }
-        )
+        agent_weight = agent_weights(records=records)
+        agent_score = agent_scores(records=records, func=costs_to_score)
+        return weighted_score(scores=agent_score, weights=agent_weight)
 
 
-def _humanness_error(costs: Costs) -> float:
+def costs_to_score(costs: Costs) -> Score:
+    """Compute score from costs.
+
+    +-------------------+--------+-----------------------------------------------------------+
+    |                   | Range  | Remarks                                                   |
+    +===================+========+===========================================================+
+    | Overall           | [0, 1] | Total score. The higher, the better.                      |
+    +-------------------+--------+-----------------------------------------------------------+
+    | DistToDestination | [0, 1] | Remaining distance to destination. The lower, the better. |
+    +-------------------+--------+-----------------------------------------------------------+
+    | Time              | [0, 1] | Time taken to complete scenario. The lower, the better.   |
+    +-------------------+--------+-----------------------------------------------------------+
+    | HumannessError    | [0, 1] | Humanness indicator. The lower, the better.               |
+    +-------------------+--------+-----------------------------------------------------------+
+    | RuleViolation     | [0, 1] | Traffic rules compliance. The lower, the better.          |
+    +-------------------+--------+-----------------------------------------------------------+
+
+    Args:
+        costs (Costs): Costs.
+
+    Returns:
+        Score: Score.
+    """
+    dist_to_destination = costs.dist_to_destination
+    humanness_error = _score_humanness_error(costs=costs)
+    rule_violation = score_rule_violation(costs=costs)
+    time = costs.steps
+    overall = (
+        0.25 * (1 - dist_to_destination)
+        + 0.25 * (1 - time)
+        + 0.25 * (1 - humanness_error)
+        + 0.25 * (1 - rule_violation)
+    )
+
+    return Score(
+        {
+            "overall": overall,
+            "dist_to_destination": dist_to_destination,
+            "time": time,
+            "humanness_error": humanness_error,
+            "rule_violation": rule_violation,
+        }
+    )
+
+
+def _score_humanness_error(costs: Costs) -> float:
     humanness_error = np.array([costs.comfort, costs.lane_center_offset])
     humanness_error = np.mean(humanness_error, dtype=float)
     return humanness_error
-
-
-def _rule_violation(costs: Costs) -> float:
-    rule_violation = np.array([costs.speed_limit, costs.wrong_way])
-    rule_violation = np.mean(rule_violation, dtype=float)
-    return rule_violation