+++ smarts/sstudio/scenario_builder.py
@@ -1,17 +1,15 @@
-# MIT License
-
 # Copyright (C) 2023. Huawei Technologies Co., Ltd. All rights reserved.
-
+#
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
 # in the Software without restriction, including without limitation the rights
 # to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
-
+#
 # The above copyright notice and this permission notice shall be included in
 # all copies or substantial portions of the Software.
-
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 # FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,55 +18,28 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from dataclasses import dataclass
+import argparse
+
+from smarts.core import seed as smarts_seed
 
 
-@dataclass(frozen=True)
-class Costs:
-    """Performance cost values."""
+def _build(scenario_py: str, seed: int):
+    # Make sure all the seed values are consistent before running the scenario script
+    smarts_seed(seed)
 
-    collisions: int = 0
-    comfort: float = 0
-    dist_to_destination: float = 0
-    dist_to_obstacles: float = 0
-    jerk_linear: float = 0
-    lane_center_offset: float = 0
-    off_road: int = 0
-    speed_limit: float = 0
-    steps: float = 0
-    vehicle_gap: float = 0
-    wrong_way: float = 0
+    # Execute the scenario script, using the current globals that were set by the seed value
+    with open(scenario_py, "rb") as source_file:
+        code = compile(source_file.read(), scenario_py, "exec")
+        exec(code, {"__file__": scenario_py})
 
 
-@dataclass(frozen=True)
-class Counts:
-    """Performance count values."""
-
-    goals: int = 0
-    """ Number of episodes completed successfully by achieving the goal.
-    """
-    episodes: int = 0
-    """ Number of episodes traversed.
-    """
-    steps: int = 0
-    """ Sum of steps taken over all episodes.
-    """
-
-
-@dataclass(frozen=True)
-class Metadata:
-    """Metadata of the record."""
-
-    difficulty: float = 1
-    """Task difficulty value.
-    """
-
-
-@dataclass
-class Record:
-    """Stores an agent's performance-cost, performance-count, and
-    performance-metadata values."""
-
-    costs: Costs
-    counts: Counts
-    metadata: Metadata
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument("scenario_py", help="Path to the scenario.py file.", type=str)
+    parser.add_argument(
+        "seed",
+        help="Seed that will be set before executing the scenario script.",
+        type=int,
+    )
+    args = parser.parse_args()
+    _build(args.scenario_py, args.seed)